Standardize tweet rendering with helper function
What: Use a helper function to generate tweets whether they are for
the home page or for a user timeline.

Why: It's modulary baby! Also, the reality is that I had user
timeline and home page rendering tweets with different code and
they got out of sync and that was uglier than ugly. Tweets should
look the same no matter what page they go to anyway.