Render user-timeline inline and show back link
What: Render starting home view of all tweets. When user
clicks on a username link, it'll hide the home view and then
will render all the tweets for that user only. Also at the top
of the user-timeline, there is a back link but it doesn't do
anything yet.

Why: Need to let the user switch between viewing main home view
and the user-timeline view for a given username.

Still more work to do.