Go back and forth between home and user-timeline
What: You can click on any username in the home view and it
will take you to their timeline view. then you can click the
back button from the timeline view and it will return you to
the home view.

Whenever you go back to the user-timeline it will be all their
latest tweets. But when you go back to home view it will be
the original tweets from when first loaded.

Why: Need to be able to go back and forth as much as you want
between home and user-timeline.

Still need to refactor a lot and make the home view fresh when
someone returns to it instead of stale though.